requirements.txt
annotated-doc==0.0.4
langchain-openai>=0.3.0
langgraph>=0.2.0
openai>=1.0.0
pydantic-settings>=2.0.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
pyyaml>=6.0
